NR-250158 - Clients in latest 15.5.0 version of extension are looping on /no-auth/provider-refresh/newrelic - Move NR specific interceptor outside of fetchCore.
