+++ shared/agent/src/system/fetchCore.ts
@@ -1,17 +1,22 @@
 import { isEmpty } from "lodash";
-import { errors, fetch, Headers, Request, RequestInfo, RequestInit, Response } from "undici";
+import { errors, fetch, Request, RequestInfo, RequestInit, Response } from "undici";
 import { Logger } from "../logger";
 import { Functions } from "./function";
 import { handleLimit, InternalRateError } from "../rateLimits";
-import { tokenHolder } from "../providers/newrelic/TokenHolder";
-import { CSAccessTokenType } from "@codestream/protocols/api";
-import { SessionContainer } from "../container";
-import { Mutex } from "async-mutex";
 
 export interface ExtraRequestInit extends RequestInit {
 	timeout?: number;
 	skipInterceptors?: boolean;
 }
+
+export type ResponseInterceptor = (
+	resp: Response,
+	init: ExtraRequestInit,
+	triedRefresh: boolean,
+	loggingPrefix: string
+) => Promise<InterceptorResponse>;
+
+export type InterceptorResponse = "retry" | "abort" | "continue";
 
 const noLogRetries = [
 	"UND_ERR_CONNECT_TIMEOUT",
@@ -24,12 +29,8 @@ const noLogRetries = [
 	"ECONNRESET",
 ];
 
-type InterceptorResponse = "retry" | "abort" | "continue";
-
-// export type ResponseInterceptor = (response: Response) => void;
-
 export class FetchCore {
-	private _refreshLock = new Mutex();
+	constructor(private readonly responseInterceptor?: ResponseInterceptor) {}
 
 	private isUndiciError(error: unknown): error is errors.UndiciError {
 		const possible = error as errors.UndiciError;
@@ -101,12 +102,15 @@ export class FetchCore {
 			}, init.timeout ?? 30000);
 			init.signal = controller.signal;
 			const resp = await fetch(url, init);
-			const interceptorResponse = await this.handleRefreshInterceptor(
-				resp,
-				init,
-				triedRefresh,
-				loggingPrefix
-			);
+			let interceptorResponse: InterceptorResponse = "continue";
+			if (this.responseInterceptor) {
+				interceptorResponse = await this.responseInterceptor(
+					resp,
+					init,
+					triedRefresh,
+					loggingPrefix
+				);
+			}
 			if (interceptorResponse === "abort") {
 				// No retry
 				return [resp, count];
@@ -132,7 +136,6 @@ export class FetchCore {
 						if (init.signal) {
 							delete init.signal;
 						}
-						// Use unmodified init
 						return this.fetchCore(count, url, init, triedRefresh);
 					}
 				}
@@ -178,103 +181,6 @@ export class FetchCore {
 			if (init.signal) {
 				delete init.signal;
 			}
-		}
-	}
-
-	private async handleRefreshInterceptor(
-		resp: Response,
-		init: ExtraRequestInit,
-		triedRefresh: boolean,
-		loggingPrefix: string
-	): Promise<InterceptorResponse> {
-		// !SessionContainer.instance().session.api.usingServiceGatewayAuth - probably not needed and
-		// at initial bootstrap it's hard to get
-		if (init.skipInterceptors || triedRefresh) {
-			return "continue";
-		}
-		const refreshToken = tokenHolder.refreshToken;
-		const isHeaders = init?.headers instanceof Headers;
-		if (!isHeaders) {
-			init.headers = new Headers(init.headers);
-		}
-		if (!resp.ok && resp.status === 403 && refreshToken) {
-			const resp2 = resp.clone();
-			const textData = await resp2.text(); // JSON response if from api-server, large HTML text blob if from service gateway
-			const isTokenExpired = textData.includes("token expired");
-			const tokenExpiredSource = this.getSource(textData, loggingPrefix);
-			if (isTokenExpired) {
-				Logger.log(`${loggingPrefix} Handling expired token from: ${tokenExpiredSource}`);
-				if (this._refreshLock.isLocked()) {
-					Logger.log(`${loggingPrefix} Waiting for already running refresh token`);
-					await this._refreshLock.waitForUnlock();
-					// TODO how to handle errors here - store in local class variable?
-					const accessToken = tokenHolder.accessToken;
-					const tokenType = tokenHolder.tokenType;
-					if (accessToken && tokenType) {
-						if (init?.headers instanceof Headers) {
-							if (tokenType === CSAccessTokenType.ACCESS_TOKEN) {
-								init.headers.set("x-access-token", accessToken);
-							} else {
-								init.headers.set("x-id-token", accessToken);
-							}
-						}
-					}
-					Logger.log(`${loggingPrefix} Refresh token wait completed`);
-					return "retry";
-				}
-				const result = this._refreshLock.runExclusive(async () => {
-					Logger.log(`${loggingPrefix} Token was found to be expired, attempting to refresh...`);
-					return await this.tokenRefresh(refreshToken, init, loggingPrefix);
-				});
-				return result;
-			}
-		}
-		return "continue"; // Not a 403 error or not token expired so let existing fetchCore logic run
-	}
-
-	private getSource(textData: string, loggingPrefix: string) {
-		// Confirmed nerdgraph and vulnerabilities reset api look the same (bith service gateway)
-		const isServiceGatewayTokenExpired = textData.includes(
-			"<title>403 Sorry – You've reached an error on New Relic</title>"
-		);
-		const isApiServerTokenExpired = textData.includes("service gateway: access token expired");
-
-		const tokenExpiredSource = isServiceGatewayTokenExpired
-			? "SG"
-			: isApiServerTokenExpired
-			? "api-server"
-			: "unknown";
-		if (tokenExpiredSource === "unknown") {
-			// TODO textData might be too big to log
-			Logger.log(`${loggingPrefix} unknown 403 error`, textData);
-		}
-		return tokenExpiredSource;
-	}
-
-	private async tokenRefresh(
-		refreshToken: string,
-		init: ExtraRequestInit,
-		loggingPrefix: string
-	): Promise<InterceptorResponse> {
-		try {
-			const tokenInfo = await SessionContainer.instance().session.api.refreshNewRelicToken(
-				refreshToken
-			);
-			Logger.log(
-				`${loggingPrefix} NR access token successfully refreshed, trying request again...`
-			);
-			if (init?.headers instanceof Headers) {
-				if (tokenInfo.tokenType === CSAccessTokenType.ACCESS_TOKEN) {
-					init.headers.set("x-access-token", tokenInfo.accessToken);
-				} else {
-					init.headers.set("x-id-token", tokenInfo.accessToken);
-				}
-			}
-			return "retry";
-		} catch (ex) {
-			Logger.warn(`${loggingPrefix} Exception thrown refreshing NR access token:`, ex);
-			return "abort";
-			// allow the original (failed) flow to continue, more meaningful than throwing an exception on refresh
 		}
 	}
 }
